# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gambit API",
    "version": "0.1.0",
    "description": "Gambit REST API — identity, sessions, seeks, ratings, and games.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    }
  },
  "paths": {
...
    "/v1/users/{handle}": {
      "get": {
        "summary": "Get a public profile with ratings",
        "tags": [
          "users"
        ],
        "operationId": "get_v1_users_by_handle",
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "description": "User handle",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWS (HS256)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "PublicUser": {
        "type": "object",
        "required": [
          "id",
          "handle",
          "country",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "handle": {
            "type": "string"
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RatingView": {
        "type": "object",
        "required": [
          "variant",
          "rating",
          "rd",
          "vol",
          "updatedAt"
        ],
        "properties": {
          "variant": {
            "type": "string",
            "enum": [
              "standard",
              "chess960",
              "kingofthehill",
              "atomic",
              "crazyhouse",
              "threecheck",
              "horde",
              "racingkings"
            ]
          },
          "rating": {
            "type": "number"
          },
          "rd": {
            "type": "number"
          },
          "vol": {
            "type": "number"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "required": [
          "user",
          "ratings"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/PublicUser"
          },
          "ratings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RatingView"
            }
          }
        }
      },
...
    }
  }
}